+++ packages/sdk/react-client-testing/src/testing.ts
@@ -1,0 +1,11 @@
+//
+// Copyright 2022 DXOS.org
+//
+
+import { TestType } from './builders';
+
+const format = (text: string) => text.replace(/\)\./g, ')\n  .');
+
+export const defaultSelectionText = format(
+  `select().filter({ type: '${TestType.Org}' }).children().filter({ type: '${TestType.Project}' })`
+);